# Cleanup code and improve the decoding performance of the Map