--- a/modules/core/src/main/scala/orcus/codec/Decoder.scala
+++ b/modules/core/src/main/scala/orcus/codec/Decoder.scala
@@ -1,10 +1,7 @@
 package orcus.codec
 
-import java.{util => ju}
-
-import cats.{Eval, Foldable}
+import cats.Eval
 import cats.instances.either._
-import cats.syntax.foldable._
 import org.apache.hadoop.hbase.client.Result
 import org.apache.hadoop.hbase.util.Bytes
 import shapeless.labelled._
@@ -56,30 +53,9 @@ object Decoder extends Decoder1 {
     def apply(result: Result): Either[Throwable, A] = a
   }
 
-  private[this] implicit val foldJavaSet: Foldable[ju.Set] = new Foldable[ju.Set] {
-    def foldLeft[A, B](fa: ju.Set[A], b: B)(f: (B, A) => B): B = {
-      @tailrec def loop(iter: ju.Iterator[A], acc: B): B = {
-        if (!iter.hasNext)
-          acc
-        else
-          loop(iter, f(acc, iter.next()))
-      }
-      loop(fa.iterator(), b)
-    }
-
-    def foldRight[A, B](fa: ju.Set[A], lb: Eval[B])(f: (A, Eval[B]) => Eval[B]): Eval[B] = {
-      def loop(iter: ju.Iterator[A]): Eval[B] = {
-        if (!iter.hasNext)
-          lb
-        else
-          f(iter.next(), Eval.defer(loop(iter)))
-      }
-      loop(fa.iterator())
-    }
-  }
-
-  implicit def decodeMap[M[_, _] <: Map[String, V], V](
+  implicit def decodeMapLike[M[_, _] <: Map[String, V], V](
       implicit
+      K: ValueCodec[String],
       V: FamilyDecoder[V],
       cbf: CanBuildFrom[Nothing, (String, V), M[String, V]]): Decoder[M[String, V]] =
     new Decoder[M[String, V]] {
@@ -91,18 +67,26 @@ object Decoder extends Decoder1 {
         val map     = result.getMap
         if (map == null) Right(builder.result())
         else {
-          val xs = map.keySet()
-          val ys = xs.foldLeftM[Either[Throwable, ?], Out](builder) {
-            case (acc, cf) =>
-              val cf0 = Bytes.toString(cf)
-              V.apply(result.getFamilyMap(cf)) match {
-                case Right(v) =>
-                  Right(acc += cf0 -> v)
-                case Left(e) =>
-                  Left(e)
+          val keys = map.keySet().iterator()
+
+          @tailrec def loop(acc: Either[Throwable, Out]): Either[Throwable, Out] = {
+            if (!keys.hasNext) acc
+            else {
+              val key = keys.next
+              K.decode(key) match {
+                case Left(e) => Left(e)
+                case Right(k) =>
+                  V.apply(result.getFamilyMap(key)) match {
+                    case Right(v) =>
+                      loop(Right(builder += k -> v))
+                    case Left(e) =>
+                      Left(e)
+                  }
               }
+            }
           }
-          ys match {
+
+          loop(Right(builder)) match {
             case Right(b) => Right(b.result())
             case Left(e)  => Left(e)
           }
@@ -111,13 +95,13 @@ object Decoder extends Decoder1 {
     }
 }
 
-trait Decoder1 extends Decoder2 {
+private[codec] trait Decoder1 extends Decoder2 {
 
-  implicit def decodeHNil: Decoder[HNil] = new Decoder[HNil] {
+  implicit val decodeHNil: Decoder[HNil] = new Decoder[HNil] {
     def apply(result: Result): Either[Throwable, HNil] = Right(HNil)
   }
 
-  implicit def decodeLabelledHList[K <: Symbol, H, T <: HList](
+  implicit def decodeLabelledHCons[K <: Symbol, H, T <: HList](
       implicit
       K: Witness.Aux[K],
       H: FamilyDecoder[H],
@@ -138,11 +122,11 @@ trait Decoder1 extends Decoder2 {
       }
     }
 
-  implicit def decodeHCons[H <: HList, A0](implicit
-                                           gen: LabelledGeneric.Aux[A0, H],
-                                           A: Lazy[Decoder[H]]): Decoder[A0] =
-    new Decoder[A0] {
-      def apply(result: Result): Either[Throwable, A0] =
+  implicit def decodeCaseClass[H <: HList, A](implicit
+                                              gen: LabelledGeneric.Aux[A, H],
+                                              A: Lazy[Decoder[H]]): Decoder[A] =
+    new Decoder[A] {
+      def apply(result: Result): Either[Throwable, A] =
         A.value(result) match {
           case Right(v) => Right(gen.from(v))
           case Left(e)  => Left(e)
@@ -150,14 +134,14 @@ trait Decoder1 extends Decoder2 {
     }
 }
 
-trait Decoder2 {
+private[codec] trait Decoder2 {
 
   implicit def decodeOption[A](implicit A: Decoder[A]): Decoder[Option[A]] =
     new Decoder[Option[A]] {
       def apply(result: Result): Either[Throwable, Option[A]] = {
         if (result.isEmpty) Right(None)
         else
-          A.apply(result) match {
+          A(result) match {
             case Right(v) => Right(Some(v))
             case Left(e)  => Left(e)
           }