--- a/modules/core/src/main/scala/orcus/codec/PutEncoder.scala
+++ b/modules/core/src/main/scala/orcus/codec/PutEncoder.scala
@@ -13,7 +13,7 @@ object PutEncoder extends PutEncoder1 {
 
   def apply[A](implicit A: PutEncoder[A]): PutEncoder[A] = A
 
-  implicit def mapPutEncoder[K, V](
+  implicit def encodeMap[K, V](
       implicit
       K: ValueCodec[K],
       V: PutFamilyEncoder[V]
@@ -21,20 +21,20 @@ object PutEncoder extends PutEncoder1 {
     def apply(acc: Put, a: Map[K, V]): Put = {
       a.foreach {
         case (k, v) =>
-          V(acc, K.encode(Option(k)), v)
+          V(acc, K.encode(k), v)
       }
       acc
     }
   }
 }
 
-trait PutEncoder1 {
+private[codec] trait PutEncoder1 {
 
-  implicit val hnilPutEncoder: PutEncoder[HNil] = new PutEncoder[HNil] {
+  implicit val encodeHNil: PutEncoder[HNil] = new PutEncoder[HNil] {
     def apply(acc: Put, a: HNil): Put = acc
   }
 
-  implicit def labelledHConsPutEncoder[K <: Symbol, H, T <: HList](
+  implicit def encodeLabelledHCons[K <: Symbol, H, T <: HList](
       implicit
       K: Witness.Aux[K],
       H: PutFamilyEncoder[H],
@@ -47,11 +47,9 @@ trait PutEncoder1 {
     }
   }
 
-  implicit def caseClassPutEncoder[A, R](
-      implicit
-      gen: LabelledGeneric.Aux[A, R],
-      R: Lazy[PutEncoder[R]]
-  ): PutEncoder[A] = new PutEncoder[A] {
+  implicit def encodeCaseClass[A, R](implicit
+                                     gen: LabelledGeneric.Aux[A, R],
+                                     R: Lazy[PutEncoder[R]]): PutEncoder[A] = new PutEncoder[A] {
     def apply(acc: Put, a: A): Put = {
       R.value(acc, gen.to(a))
     }