--- a/modules/core/src/main/scala/orcus/codec/package.scala
+++ b/modules/core/src/main/scala/orcus/codec/package.scala
@@ -1,8 +1,0 @@
-package orcus
-
-package object codec {
-
-  implicit class ArrayOps private[orcus] (val a: Option[Array[Byte]]) extends AnyVal {
-    def orEmpty: Array[Byte] = a.getOrElse(Array.emptyByteArray)
-  }
-}