current master

```
[info] Benchmark                                    (size)   Mode  Cnt       Score      Error  Units
[info] DecoderBench.decodeSelf                          10  thrpt   20  361210.915 ± 6105.123  ops/s
[info] DecoderBench.decodeSelf                          30  thrpt   20  103226.093 ± 1403.922  ops/s
[info] DecoderBench.decodeToCaseClass                   10  thrpt   20  328102.225 ± 6773.216  ops/s
[info] DecoderBench.decodeToCaseClass                   30  thrpt   20  101035.228 ± 1049.479  ops/s
[info] DecoderBench.decodeToCaseClassCachedDecoder      10  thrpt   20  353435.104 ± 6272.487  ops/s
[info] DecoderBench.decodeToCaseClassCachedDecoder      30  thrpt   20  105248.537 ± 1794.158  ops/s
[info] DecoderBench.decodeToMap                         10  thrpt   20  257109.247 ± 9525.354  ops/s
[info] DecoderBench.decodeToMap                         30  thrpt   20   87897.808 ± 1161.379  ops/s
```

this PR

```
[info] Benchmark                                    (size)   Mode  Cnt       Score       Error  Units
[info] DecoderBench.decodeSelf                          10  thrpt   20  362397.142 ±  7316.538  ops/s
[info] DecoderBench.decodeSelf                          30  thrpt   20  104472.350 ±  1569.480  ops/s
[info] DecoderBench.decodeToCaseClass                   10  thrpt   20  323035.057 ±  5057.819  ops/s
[info] DecoderBench.decodeToCaseClass                   30  thrpt   20  100794.912 ±  1896.704  ops/s
[info] DecoderBench.decodeToCaseClassCachedDecoder      10  thrpt   20  350700.835 ± 11481.967  ops/s
[info] DecoderBench.decodeToCaseClassCachedDecoder      30  thrpt   20  105769.920 ±  1213.758  ops/s
[info] DecoderBench.decodeToMap                         10  thrpt   20  278057.669 ±  5989.611  ops/s
[info] DecoderBench.decodeToMap                         30  thrpt   20   88068.112 ±  1948.720  ops/s
```